[exp-floating-ui-v2] Move maybeGetPortalMountedModalHostElement to Floating
