+++ packages/wonder-blocks-dropdown/package.json
@@ -29,10 +29,10 @@
     "@khanacademy/wonder-blocks-cell": "workspace:*",
     "@khanacademy/wonder-blocks-clickable": "workspace:*",
     "@khanacademy/wonder-blocks-core": "workspace:*",
+    "@khanacademy/wonder-blocks-floating": "workspace:*",
     "@khanacademy/wonder-blocks-form": "workspace:*",
     "@khanacademy/wonder-blocks-icon": "workspace:*",
     "@khanacademy/wonder-blocks-icon-button": "workspace:*",
-    "@khanacademy/wonder-blocks-modal": "workspace:*",
     "@khanacademy/wonder-blocks-pill": "workspace:*",
     "@khanacademy/wonder-blocks-search-field": "workspace:*",
     "@khanacademy/wonder-blocks-styles": "workspace:*",
