+++ packages/wonder-blocks-dropdown/tsconfig-build.json
@@ -10,9 +10,9 @@
         {"path": "../wonder-blocks-cell/tsconfig-build.json"},
         {"path": "../wonder-blocks-clickable/tsconfig-build.json"},
         {"path": "../wonder-blocks-core/tsconfig-build.json"},
+        {"path": "../wonder-blocks-floating/tsconfig-build.json"},
         {"path": "../wonder-blocks-form/tsconfig-build.json"},
         {"path": "../wonder-blocks-icon/tsconfig-build.json"},
-        {"path": "../wonder-blocks-modal/tsconfig-build.json"},
         {"path": "../wonder-blocks-pill/tsconfig-build.json"},
         {"path": "../wonder-blocks-search-field/tsconfig-build.json"},
         {"path": "../wonder-blocks-styles/tsconfig-build.json"},
