+++ packages/wonder-blocks-floating/src/index.ts
@@ -1,1 +1,2 @@
 export {default as Floating} from "./components/floating";
+export {default as maybeGetPortalMountedModalHostElement} from "./util/maybe-get-portal-mounted-modal-host-element";
