+++ packages/wonder-blocks-floating/src/util/constants.ts
@@ -1,0 +1,6 @@
+/**
+ * The attribute used to identify a modal launcher portal.
+ */
+const ModalLauncherPortalAttributeName = "data-modal-launcher-portal";
+
+export {ModalLauncherPortalAttributeName};
