+++ packages/wonder-blocks-tooltip/src/index.ts
@@ -2,9 +2,7 @@ import type {Placement, PopperElementProps} from "./util/types";
 
 import Tooltip from "./components/tooltip";
 import TooltipContent from "./components/tooltip-content";
-import TooltipPopper from "./components/tooltip-popper";
-import TooltipTail from "./components/tooltip-tail";
 
-export {Tooltip as default, TooltipContent, TooltipPopper, TooltipTail};
+export {Tooltip as default, TooltipContent};
 
 export type {Placement, PopperElementProps};
